+++ cmp-worker-kmp/src/jsMain/kotlin/io/github/mobilebytelabs/worker/BackgroundCapabilities.js.kt
@@ -1,0 +1,6 @@
+package io.github.mobilebytelabs.worker
+
+actual fun platformBackgroundCapabilities() = BackgroundCapabilities(
+    supportsPersistence = true,
+    supportsOsScheduling = false,
+)