+++ cmp-worker-kmp/src/wasmJsMain/kotlin/io/github/mobilebytelabs/worker/BackgroundCapabilities.wasmJs.kt
@@ -1,0 +1,6 @@
+package io.github.mobilebytelabs.worker
+
+actual fun platformBackgroundCapabilities() = BackgroundCapabilities(
+    supportsPersistence = false,
+    supportsOsScheduling = false,
+)